+++ CHANGES.rst
@@ -4,7 +4,9 @@
 - Do not enumerate `search_path` with external schemas (`Issue #120
   <https://github.com/sqlalchemy-redshift/sqlalchemy-redshift/pull/120>`_)
 - Return constraint name from get_pk_constraint and get_foreign_keys
-
+- Update certificate bundle for transition to ACM certs
+  https://docs.aws.amazon.com/redshift/latest/mgmt/connecting-transitioning-to-acm-certs.html
+(`Issue #130 <https://github.com/sqlalchemy-redshift/sqlalchemy-redshift/pull/130>`_)
 
 0.6.0 (2017-05-04)
 ------------------